Competition-Solutions/Text/Lacuna Masakhane Parts of Speech Classification Challenge/#1 randomseed19/requirements.txt
numpy==1.23.5
pandas==1.5.3
torch==2.0.1+cu118
scikit-learn==1.2.2
transformers==4.31.0
datasets==2.14.4
evaluate==0.4.0
seqeval==1.2.2
accelerate==0.22.0
peft==0.5.0
bitsandbytes==0.41.1
wandb==0.15.8
tensorboard==2.12.3
python-dotenv==1.0.0
tqdm==4.66.1
sentencepiece==0.1.99
